refactor: 네이버 지도 관련 util 함수 분리
